Fix: link layer doesn't hadle client exception properly. Add: async call (inflight datagram) to retrieve AL status code and AL error code Improve: retrieve whole AL status (code + error) in one access instead of two access.

+++ example/easycat_example.cc
@@ -136,7 +136,7 @@ int main(int argc, char* argv[])
         }
         catch (std::exception const& e)
         {
-            int delta = i - last_error;
+            int64_t delta = i - last_error;
             last_error = i;
             std::cerr << e.what() << " at " << i << " delta: " << delta << std::endl;
         }
